add category to diagnostic (required vs recommended)
