+++ tools/contract-analyzer/analyzers/analyzers_test.go
@@ -63,6 +63,7 @@ func TestReferenceToOptionalAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 160, Line: 5, Column: 44},
 				},
 				Location:         testLocation,
+				Category:         "update required",
 				Message:          "reference to optional will return optional reference",
 				SecondaryMessage: "replace with '&String?'",
 			},
@@ -96,6 +97,7 @@ func TestDeprecatedKeyFunctionsAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 119, Line: 4, Column: 37},
 				},
 				Location:         testLocation,
+				Category:         "update recommended",
 				Message:          "deprecated function 'addPublicKey' will get removed",
 				SecondaryMessage: "replace with 'keys.add'",
 			},
@@ -105,6 +107,7 @@ func TestDeprecatedKeyFunctionsAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 165, Line: 5, Column: 40},
 				},
 				Location:         testLocation,
+				Category:         "update recommended",
 				Message:          "deprecated function 'removePublicKey' will get removed",
 				SecondaryMessage: "replace with 'keys.revoke'",
 			},
@@ -137,6 +140,7 @@ func TestNumberSupertypeBinaryOperationsAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 114, Line: 4, Column: 30},
 				},
 				Location: testLocation,
+				Category: "update required",
 				Message:  "arithmetic operations on number supertypes will get removed",
 			},
 		},
@@ -167,16 +171,20 @@ func TestParameterListMissingCommasAnalyzer(t *testing.T) {
 					StartPos: ast.Position{Offset: 64, Line: 3, Column: 33},
 					EndPos:   ast.Position{Offset: 64, Line: 3, Column: 33},
 				},
-				Location: testLocation,
-				Message:  "missing comma between parameters",
+				Location:         testLocation,
+				Category:         "update required",
+				Message:          "missing comma between parameters",
+				SecondaryMessage: "insert missing comma here",
 			},
 			{
 				Range: ast.Range{
 					StartPos: ast.Position{Offset: 108, Line: 4, Column: 29},
 					EndPos:   ast.Position{Offset: 108, Line: 4, Column: 29},
 				},
-				Location: testLocation,
-				Message:  "missing comma between parameters",
+				Location:         testLocation,
+				Category:         "update required",
+				Message:          "missing comma between parameters",
+				SecondaryMessage: "insert missing comma here",
 			},
 		},
 		diagnostics,
@@ -211,6 +219,7 @@ func TestSupertypeInferenceAnalyzer(t *testing.T) {
 						EndPos:   ast.Position{Offset: 148, Line: 6, Column: 52},
 					},
 					Location:         testLocation,
+					Category:         "check required",
 					Message:          "type inference for arrays will change",
 					SecondaryMessage: "ensure the newly inferred type is correct",
 				},
@@ -244,6 +253,7 @@ func TestSupertypeInferenceAnalyzer(t *testing.T) {
 						EndPos:   ast.Position{Offset: 168, Line: 7, Column: 55},
 					},
 					Location:         testLocation,
+					Category:         "check required",
 					Message:          "type inference for dictionaries will change",
 					SecondaryMessage: "ensure the newly inferred type is correct",
 				},
@@ -277,6 +287,7 @@ func TestSupertypeInferenceAnalyzer(t *testing.T) {
 						EndPos:   ast.Position{Offset: 154, Line: 7, Column: 51},
 					},
 					Location:         testLocation,
+					Category:         "check required",
 					Message:          "type inference for conditionals / ternary operations will change",
 					SecondaryMessage: "ensure the newly inferred type is correct",
 				},
@@ -349,6 +360,7 @@ func TestExternalMutationAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 891, Line: 31, Column: 26},
 				},
 				Location:         testLocation,
+				Category:         "update required",
 				Message:          "external mutation of non-settable public container-typed field will get disallowed",
 				SecondaryMessage: "add setter function for field, or change field access to pub(set)",
 			},
@@ -358,6 +370,7 @@ func TestExternalMutationAnalyzer(t *testing.T) {
 					EndPos:   ast.Position{Offset: 1135, Line: 39, Column: 26},
 				},
 				Location:         testLocation,
+				Category:         "update required",
 				Message:          "external mutation of non-settable public container-typed field will get disallowed",
 				SecondaryMessage: "add setter function for field, or change field access to pub(set)",
 			},
